refactor: backend functions
