+++ apps/web/src/actions/unirefund/TravellerService/post-actions.ts
@@ -3,15 +3,19 @@ import type {
   PostApiTravellerServiceTravellersByIdCreatePersonalIdentificationData,
   PostApiTravellerServiceTravellersData,
 } from "@ayasofyazilim/saas/TravellerService";
-import { structuredError, structuredResponse } from "../../../lib";
-import { getApiRequests } from "../../api-requests";
+import {
+  getTravellersServiceClient,
+  structuredError,
+  structuredResponse,
+} from "../../../lib";
 
 export async function postTravellerApi(
   data: PostApiTravellerServiceTravellersData,
 ) {
   try {
-    const client = await getApiRequests();
-    const response = await client.travellers.post(data);
+    const client = await getTravellersServiceClient();
+    const response =
+      await client.traveller.postApiTravellerServiceTravellers(data);
     return structuredResponse(response);
   } catch (error) {
     return structuredError(error);
@@ -22,8 +26,11 @@ export async function postTravellerIdentificationApi(
   data: PostApiTravellerServiceTravellersByIdCreatePersonalIdentificationData,
 ) {
   try {
-    const client = await getApiRequests();
-    const response = await client.travellers.postTravellerIdentification(data);
+    const client = await getTravellersServiceClient();
+    const response =
+      await client.traveller.putApiTravellerServiceTravellersByIdUpdatePersonalIdentification(
+        data,
+      );
     return structuredResponse(response);
   } catch (error) {
     return structuredError(error);
